generate a file to pass flags to ghc-7.6
Ignore-this: 33b5260c6456c348bf03b2439f9aa60c

The reason for this change is that ghc-7.6 doesn't have a cabal repl when
using the Cabal that comes with ghc. The build-type is not updated to Custom,
since normally we do have a new-enough Cabal.

darcs-hash:06bc949dba04eaedb5a430364ce5ece3ecfbb278