depend on base-orphans for instance Typeable '[] and '(:)
Ignore-this: 44c36eee01485d21f5a209cdc9582eca

When compiling the test suite with ghc-7.6, somehow base-orphans
was being brought in (via hspec?), and duplicate instances are
not allowed.  It is possible to only depend on that package
for ghc < 7.7, but it seems simpler to always bring it in.

darcs-hash:5c6c63c5b6fa44c5eb88d48be3fb691269ecaa4d